+++ back-end/api/src/main/java/com/chatop/api/dto/RentalDto.java
@@ -6,7 +6,6 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -16,8 +15,8 @@ import java.time.LocalDateTime;
 public class RentalDto {
     private Integer id;
     private String name;
-    private BigDecimal surface;
-    private BigDecimal price;
+    private Integer surface;
+    private Integer price;
     private String picture;
     private String description;
     private Integer owner_id; // On renvoie juste l'ID, pas tout l'objet User
